requirements.txt
# torch is installed separately in Dockerfile as CPU-only to save memory
fastapi
uvicorn
pandas
scikit-learn
faiss-cpu
sentence-transformers
openai
langchain
langchain-openai
langchain-core
langchain-community
joblib
python-multipart
python-dotenv
